OpenAI/requirements.txt
langchain
langchain_openai
streamlit
python-dotenv
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
